2024/requirements.txt
beautifulsoup4>=4.12.3,<4.13
requests>=2.32.3,<2.33
python-dotenv>= 1.0.1